now we can train with sdss galaxies
